+++ src-tauri/src/profile/launch/mod.rs
@@ -5,7 +5,7 @@ use std::{
     process::Command,
 };
 
-use eyre::{bail, ensure, eyre, Context, OptionExt, Result};
+use eyre::{bail, ensure, eyre, OptionExt, Result};
 use log::{info, warn};
 use serde::{Deserialize, Serialize};
 use tauri::AppHandle;
@@ -13,11 +13,17 @@ use tokio::time::Duration;
 
 use super::ManagedGame;
 use crate::{
-    game::{Game, ModLoader, ModLoaderKind, Platform},
+    game::Game,
     logger::log_webview_err,
     prefs::{GamePrefs, Prefs},
-    util::{self, error::IoResultExt},
+    util::{
+        self,
+        fs::{Overwrite, UseLinks},
+    },
 };
+
+mod mod_loader;
+mod platform;
 
 pub mod commands;
 
@@ -61,34 +67,20 @@ impl ManagedGame {
                 Default::default()
             });
 
-        // if the game has a platform, but the setting is unset, fill it in anyway
+        // if the game has a platform but the setting is unset, fill it in
         platform = platform.or_else(|| self.game.platforms.iter().next());
 
         let mut command = match (&launch_mode, platform) {
-            (LaunchMode::Launcher, Some(Platform::Steam)) => steam_command(self.game, prefs)?,
-            (LaunchMode::Launcher, Some(Platform::EpicGames)) => {
-                let Some(epic) = &self.game.platforms.epic_games else {
-                    bail!("{} is not available on Epic Games", self.game.name)
-                };
-
-                let url = format!(
-                    "com.epicgames.launcher://apps/{}?action=launch&silent=true",
-                    epic.identifier.unwrap_or(self.game.name)
-                );
-
-                info!("launching from Epic Games with URL {}", url);
-
-                open::commands(url)
-                    .into_iter()
-                    .next()
-                    .ok_or_eyre("open returned no commands to try")?
+            (LaunchMode::Launcher, Some(platform)) => {
+                platform::launch_command(platform, self.game, prefs).transpose()
             }
-            (_, _) => Command::new(exe_path(game_dir)?),
-        };
+            _ => None,
+        }
+        .unwrap_or_else(|| exe_path(game_dir).map(Command::new))?;
 
         let profile = self.active_profile();
 
-        add_loader_args(&mut command, &profile.path, &self.game.mod_loader)?;
+        mod_loader::add_args(&mut command, &profile.path, &self.game.mod_loader)?;
 
         if let Some(custom_args) = custom_args {
             command.args(custom_args);
@@ -111,7 +103,10 @@ impl ManagedGame {
             .path
             .read_dir()?
             .filter_map(|entry| entry.ok())
-            .filter(|entry| entry.file_type().is_ok_and(|ty| ty.is_file()))
+            .filter(|entry| {
+                return entry.file_type().is_ok_and(|ty| ty.is_file())
+                    || entry.file_name() == "dotnet"; // bepinex il2cpp libraries
+            })
             .filter(|entry| {
                 let name = entry.file_name();
                 EXCLUDES.iter().all(|exclude| name != *exclude)
@@ -122,176 +117,21 @@ impl ManagedGame {
                 "copying {} to game directory",
                 file.file_name().to_string_lossy()
             );
-            fs::copy(file.path(), game_dir.join(file.file_name()))?;
+
+            if file.file_type().is_ok_and(|ty| ty.is_file()) {
+                fs::copy(file.path(), game_dir.join(file.file_name()))?;
+            } else {
+                util::fs::copy_dir(
+                    &file.path(),
+                    &game_dir.join(file.file_name()),
+                    Overwrite::Yes,
+                    UseLinks::No,
+                )?;
+            }
         }
 
         Ok(())
     }
-}
-
-fn steam_command(game: Game, prefs: &Prefs) -> Result<Command> {
-    let Some(steam) = &game.platforms.steam else {
-        bail!("{} is not available on Steam", game.name)
-    };
-
-    let steam_path = prefs
-        .steam_exe_path
-        .as_ref()
-        .ok_or_eyre("steam executable path not set")?;
-
-    ensure!(
-        steam_path.exists(),
-        "steam executable not found at {}",
-        steam_path.display()
-    );
-
-    let mut command = Command::new(steam_path);
-    command.arg("-applaunch").arg(steam.id.to_string());
-
-    Ok(command)
-}
-
-fn game_dir(game: Game, prefs: &Prefs) -> Result<PathBuf> {
-    let game_prefs = prefs.game_prefs.get(&*game.slug);
-
-    let path = if let Some(GamePrefs {
-        dir_override: Some(path),
-        ..
-    }) = game_prefs
-    {
-        info!("using game path override at {}", path.display());
-        path.to_path_buf()
-    } else {
-        let platform = game_prefs
-            .and_then(|prefs| prefs.platform)
-            .or_else(|| game.platforms.iter().next());
-
-        match platform {
-            Some(Platform::Steam) => steam_game_dir(game, prefs)?,
-            #[cfg(windows)]
-            Some(Platform::XboxStore) => xbox_game_dir(game)?,
-            #[cfg(windows)]
-            Some(Platform::EpicGames) => epic_game_dir(game)?,
-            _ => bail!("game directory not found - you may need to specify it in the settings"),
-        }
-    };
-
-    ensure!(
-        path.exists(),
-        "game directory does not exist, please check your settings (expected at {})",
-        path.display()
-    );
-
-    Ok(path)
-}
-
-fn steam_game_dir(game: Game, prefs: &Prefs) -> Result<PathBuf> {
-    let Some(steam) = &game.platforms.steam else {
-        bail!("{} is not available on Steam", game.name)
-    };
-
-    let mut path = prefs
-        .steam_library_dir
-        .as_ref()
-        .ok_or_eyre("steam library directory not set")?
-        .to_path_buf();
-
-    if !path.ends_with("common") {
-        if !path.ends_with("steamapps") {
-            path.push("steamapps");
-        }
-
-        path.push("common");
-    }
-
-    info!(
-        "using {} path from steam library at {}",
-        game.slug,
-        path.display()
-    );
-
-    path.push(steam.dir_name.unwrap_or(game.name));
-
-    Ok(path)
-}
-
-#[cfg(windows)]
-fn xbox_game_dir(game: Game) -> Result<PathBuf> {
-    let Some(xbox) = &game.platforms.xbox_store else {
-        bail!("{} is not available on Xbox Store", game.name)
-    };
-
-    let name = xbox.identifier.unwrap_or(game.name);
-    let mut query = Command::new("powershell.exe");
-    query.args([
-        "get-appxpackage",
-        "-Name",
-        name,
-        "|",
-        "select",
-        "-expand",
-        "InstallLocation",
-    ]);
-
-    info!("querying path for {} with command {:?}", game.slug, query);
-
-    let out = query.output()?;
-
-    ensure!(
-        out.status.success(),
-        "query returned with error code {}",
-        out.status.code().unwrap_or(-1)
-    );
-
-    let str = String::from_utf8(out.stdout).context("query returned invalid UTF-8")?;
-
-    Ok(PathBuf::from(str))
-}
-
-#[cfg(windows)]
-fn epic_game_dir(game: &crate::game::GameData<'_>) -> Result<PathBuf, eyre::Error> {
-    let Some(epic) = &game.platforms.epic_games else {
-        bail!("{} is not available on Epic Games", game.name)
-    };
-
-    let name = epic.identifier.unwrap_or(game.name);
-    let dat_path: PathBuf =
-        PathBuf::from("C:/ProgramData/Epic/UnrealEngineLauncher/LauncherInstalled.dat");
-
-    #[derive(Debug, Deserialize)]
-    #[serde(rename_all = "PascalCase")]
-    struct ListItem {
-        install_location: PathBuf,
-        app_name: String,
-    }
-
-    info!(
-        "reading Epic Games installations from {}",
-        dat_path.display()
-    );
-
-    let list: Vec<ListItem> =
-        util::fs::read_json(dat_path).context("failed to read LauncherInstalled.dat file")?;
-
-    list.into_iter()
-        .find(|item| item.app_name == name)
-        .map(|item| item.install_location)
-        .ok_or_eyre("could not find entry in the list of installed games")
-}
-
-fn exe_path(game_dir: &Path) -> Result<PathBuf> {
-    game_dir
-        .read_dir()?
-        .filter_map(Result::ok)
-        .find(|entry| {
-            let file_name = PathBuf::from(entry.file_name());
-            let extension = file_name.extension().and_then(|ext| ext.to_str());
-
-            matches!(extension, Some("exe" | "sh"))
-                && !file_name.to_string_lossy().contains("UnityCrashHandler")
-        })
-        .map(|entry| entry.path())
-        .ok_or_eyre("game executable not found")
 }
 
 fn do_launch(mut command: Command, app: AppHandle, mode: LaunchMode) -> Result<()> {
@@ -322,142 +162,49 @@ fn do_launch(mut command: Command, app: AppHandle, mode: LaunchMode) -> Result<(
     Ok(())
 }
 
-fn add_loader_args(
-    command: &mut Command,
-    profile_dir: &Path,
-    mod_loader: &ModLoader,
-) -> Result<()> {
-    match &mod_loader.kind {
-        ModLoaderKind::BepInEx { .. } => add_bepinex_args(command, profile_dir),
-        ModLoaderKind::MelonLoader { .. } => add_melon_loader_args(command, profile_dir),
-        ModLoaderKind::Northstar {} => add_northstar_args(command, profile_dir),
-        ModLoaderKind::GDWeave {} => add_gd_weave_args(command, profile_dir),
-        ModLoaderKind::Shimloader {} => add_shimloader_args(command, profile_dir),
-        ModLoaderKind::Lovely {} => add_lovely_args(command, profile_dir),
-        ModLoaderKind::ReturnOfModding { .. } => add_return_of_modding_args(command, profile_dir),
-    }
-}
+fn game_dir(game: Game, prefs: &Prefs) -> Result<PathBuf> {
+    let game_prefs = prefs.game_prefs.get(&*game.slug);
 
-fn add_bepinex_args(command: &mut Command, profile_dir: &Path) -> Result<()> {
-    let (enable_prefix, target_prefix) = doorstop_args(profile_dir)?;
-    let preloader_path = bepinex_preloader_path(profile_dir)?;
-
-    command
-        .args([enable_prefix, "true", target_prefix])
-        .arg(preloader_path);
-
-    Ok(())
-}
-
-fn bepinex_preloader_path(profile_dir: &Path) -> Result<PathBuf> {
-    let mut core_dir = profile_dir.to_path_buf();
-
-    core_dir.push("BepInEx");
-    core_dir.push("core");
-
-    const PRELOADER_NAMES: &[&str] = &[
-        "BepInEx.Unity.Mono.Preloader.dll",
-        "BepInEx.Unity.IL2CPP.dll",
-        "BepInEx.Preloader.dll",
-        "BepInEx.IL2CPP.dll",
-    ];
-
-    let result = core_dir
-        .read_dir()
-        .context("failed to read BepInEx core directory. Is BepInEx installed?")?
-        .filter_map(|entry| entry.ok())
-        .find(|entry| {
-            let file_name = entry.file_name();
-            PRELOADER_NAMES.iter().any(|name| file_name == **name)
-        })
-        .ok_or_eyre("BepInEx preloader not found. Is BepInEx installed?")?
-        .path();
-
-    Ok(result)
-}
-
-fn doorstop_args(profile_dir: &Path) -> Result<(&'static str, &'static str)> {
-    let path = profile_dir.join(".doorstop_version");
-
-    let version = if path.exists() {
-        let version = fs::read_to_string(&path)
-            .fs_context("reading version file", &path)?
-            .split('.') // read only the major version number
-            .next()
-            .and_then(|str| str.parse().ok())
-            .ok_or_eyre("invalid version format")?;
-
-        info!("doorstop version read: {}", version);
-        version
+    let path = if let Some(GamePrefs {
+        dir_override: Some(path),
+        ..
+    }) = game_prefs
+    {
+        info!("using game path override at {}", path.display());
+        path.to_path_buf()
     } else {
-        warn!(".doorstop_version file is missing, defaulting to 3");
-        3
+        let platform = game_prefs
+            .and_then(|prefs| prefs.platform)
+            .or_else(|| game.platforms.iter().next());
+
+        platform::game_dir(platform, game, prefs)?
     };
 
-    match version {
-        3 => Ok(("--doorstop-enable", "--doorstop-target")),
-        4 => Ok(("--doorstop-enabled", "--doorstop-target-assembly")),
-        vers => bail!("unsupported doorstop version: {}", vers),
-    }
+    ensure!(
+        path.exists(),
+        "game directory does not exist, please check your settings (expected at {})",
+        path.display()
+    );
+
+    Ok(path)
 }
 
-fn add_melon_loader_args(command: &mut Command, profile_dir: &Path) -> Result<()> {
-    command.arg("--melonloader.basedir").arg(profile_dir);
+fn exe_path(game_dir: &Path) -> Result<PathBuf> {
+    game_dir
+        .read_dir()?
+        .filter_map(Result::ok)
+        .find(|entry| {
+            let file_name = PathBuf::from(entry.file_name());
+            let extension = file_name.extension().and_then(|ext| ext.to_str());
 
-    let agf_path = profile_dir.join("MelonLoader/Managed/Assembly-CSharp.dll");
+            let has_correct_extension = if cfg!(windows) {
+                matches!(extension, Some("exe"))
+            } else {
+                matches!(extension, Some("exe" | "sh"))
+            };
 
-    if !agf_path.exists() {
-        command.arg("--melonloader.agfregenerate");
-    }
-
-    Ok(())
-}
-
-fn add_northstar_args(command: &mut Command, profile_dir: &Path) -> Result<()> {
-    let path = profile_dir.join("R2Northstar");
-    let path = path
-        .to_str()
-        .ok_or_eyre("profile path is not valid UTF-8")?;
-
-    command.arg("-northstar").arg(format!("-profile={}", path));
-
-    Ok(())
-}
-
-fn add_gd_weave_args(command: &mut Command, profile_dir: &Path) -> Result<()> {
-    let path = profile_dir.join("GDWeave");
-    let path = path
-        .to_str()
-        .ok_or_eyre("profile path is not valid UTF-8")?;
-
-    command.arg(format!("--gdweave-folder-override={}", path));
-
-    Ok(())
-}
-
-fn add_shimloader_args(command: &mut Command, profile_dir: &Path) -> Result<()> {
-    let path = profile_dir.join("shimloader");
-
-    command
-        .arg("--mod-dir")
-        .arg(path.join("mod"))
-        .arg("--pak-dir")
-        .arg(path.join("pak"))
-        .arg("--cfg-dir")
-        .arg(path.join("cfg"));
-
-    Ok(())
-}
-
-fn add_lovely_args(command: &mut Command, profile_dir: &Path) -> Result<()> {
-    let path = profile_dir.join("mods");
-    command.arg("--mod-dir").arg(path);
-
-    Ok(())
-}
-
-fn add_return_of_modding_args(command: &mut Command, profile_dir: &Path) -> Result<()> {
-    command.arg("--rom_modding_root_folder").arg(profile_dir);
-
-    Ok(())
+            has_correct_extension && !file_name.to_string_lossy().contains("UnityCrashHandler")
+        })
+        .map(|entry| entry.path())
+        .ok_or_eyre("game executable not found")
 }